requirements.txt
PyNaCl>=1.3.0
requests>=2.22.0
urllib3>=1.25.7